+++ packages/flame/test/sprite_batch_test.dart
@@ -1,9 +1,13 @@
 import 'dart:ui';
 
+import 'package:flame/components.dart';
 import 'package:flame/sprite.dart';
+import 'package:flame_test/flame_test.dart';
 import 'package:flutter/material.dart' hide Image;
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mocktail/mocktail.dart';
+
+import '_resources/load_image.dart';
 
 class _MockImage extends Mock implements Image {}
 
@@ -56,5 +60,68 @@ void main() {
             .having((t) => t.ty, 'ty', 1),
       );
     });
+
+    const margin = 2.0;
+    const tileSize = 6.0;
+
+    testGolden(
+      'can render a batch with blend mode',
+      (game) async {
+        final spriteSheet = await loadImage('alphabet.png');
+        final spriteBatch = SpriteBatch(spriteSheet);
+
+        const source = Rect.fromLTWH(3 * tileSize, 0, tileSize, tileSize);
+
+        spriteBatch.add(
+          source: source,
+          color: Colors.redAccent,
+          offset: Vector2.all(margin),
+        );
+
+        spriteBatch.add(
+          source: source,
+          offset: Vector2(2 * margin + tileSize, margin),
+        );
+
+        game.add(
+          SpriteBatchComponent(
+            spriteBatch: spriteBatch,
+            blendMode: BlendMode.srcOver,
+          ),
+        );
+      },
+      size: Vector2(3 * margin + 2 * tileSize, 2 * margin + tileSize),
+      backgroundColor: const Color(0xFFFFFFFF),
+      goldenFile: '_goldens/sprite_batch_test_1.png',
+    );
+
+    testGolden(
+      'can render a batch without blend mode',
+      (game) async {
+        final spriteSheet = await loadImage('alphabet.png');
+        final spriteBatch = SpriteBatch(spriteSheet);
+
+        const source = Rect.fromLTWH(3 * tileSize, 0, tileSize, tileSize);
+
+        spriteBatch.add(
+          source: source,
+          offset: Vector2.all(margin),
+        );
+
+        spriteBatch.add(
+          source: source,
+          offset: Vector2(2 * margin + tileSize, margin),
+        );
+
+        game.add(
+          SpriteBatchComponent(
+            spriteBatch: spriteBatch,
+          ),
+        );
+      },
+      size: Vector2(3 * margin + 2 * tileSize, 2 * margin + tileSize),
+      backgroundColor: const Color(0xFFFFFFFF),
+      goldenFile: '_goldens/sprite_batch_test_2.png',
+    );
   });
 }