+++ packages/flame_fire_atlas/test/flame_fire_atlas_test.dart
@@ -61,6 +61,32 @@ void main() {
       expect(atlas.id, 'cave_ace');
     });
 
+    test('image returns the loaded image', () async {
+      final atlas = await _readTestAtlas();
+      final image = atlas.image;
+      expect(image, isA<Image>());
+    });
+
+    test('image throws when the image is not loaded', () async {
+      final atlas = FireAtlas(
+        id: '',
+        tileWidth: 0,
+        tileHeight: 0,
+        imageData: '',
+      );
+
+      expect(
+        () => atlas.image,
+        throwsA(
+          isA<Exception>().having(
+            (e) => e.toString(),
+            'toString',
+            'Exception: Atlas is not loaded yet, call "load" before using it',
+          ),
+        ),
+      );
+    });
+
     test('can load the asset using the global assets/images', () async {
       final assetsMock = _AssetsCacheMock();