+++ packages/flame_behavior_tree/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.1.3+2
+
+ - Update a dependency to the latest release.
+
 ## 0.1.3+1
 
  - Update a dependency to the latest release.