+++ packages/flame_kenney_xml/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.1.1+2
+
+ - Update a dependency to the latest release.
+
 ## 0.1.1+1
 
  - Update a dependency to the latest release.